added computer strategy
changed that game so that the game object checks the validity of the cards rather than the player. This allows for a very basic computerPlayer that doesn't even try to choose a valid card. Implemented an extension to ComputerPlayer called RandomComputerPlayer that uses the random valid card strategy. This will give students two different exmaples of strategies. This also makes the game slightly more fun as human  players have to be careful not to accidently choose the wrong card.