revisions from feedback validate_card_choice, tqdm, wild card bug
Added a function to return valid card choices from a hand in the player class, made the pick card functions raise an exception in the player base class. Fixed bug where a wild card drawn as the first card would not have a color associated with it. Wrote a command line interface for the game. Added tqmd progress bar to the test file. Added requirements.txt file. Created documentation in README for installation and features of the game.